autorole: fix issue in DMs where message.guild is null

+++ cogs/autorole.py
@@ -20,18 +20,19 @@ class Autorole(commands.Cog):
     async def on_message(self, message):
         # avoid a stupid race condition, apparently the bot checks its own messages. What?
         if message.author.id != self.bot.user.id:
-            if message.guild.id == self.bot.settings['GUILD'] and message.channel.id == self.bot.settings['WELCOME']:
-                await message.delete()
-                if message.content.lower() == "g":
-                    return await message.author.add_roles(message.guild.get_role(self.bot.settings['SETG_ROLE']))
-                elif message.content.lower() == "h":
-                    return await message.author.add_roles(message.guild.get_role(self.bot.settings['SETH_ROLE']))
-                elif message.content.lower() == "j":
-                    return await message.author.add_roles(message.guild.get_role(self.bot.settings['SETJ_ROLE']))
-                elif message.content.lower() == "f":
-                    return await message.author.add_roles(message.guild.get_role(self.bot.settings['DEFAULT_ROLE']))
-                else:
-                    return await message.channel.send(content="Invalid class. Try again.", delete_after=10)
+            if message.guild is not None:
+                if message.guild.id == self.bot.settings['GUILD'] and message.channel.id == self.bot.settings['WELCOME']:
+                    await message.delete()
+                    if message.content.lower() == "g":
+                        return await message.author.add_roles(message.guild.get_role(self.bot.settings['SETG_ROLE']))
+                    elif message.content.lower() == "h":
+                        return await message.author.add_roles(message.guild.get_role(self.bot.settings['SETH_ROLE']))
+                    elif message.content.lower() == "j":
+                        return await message.author.add_roles(message.guild.get_role(self.bot.settings['SETJ_ROLE']))
+                    elif message.content.lower() == "f":
+                        return await message.author.add_roles(message.guild.get_role(self.bot.settings['DEFAULT_ROLE']))
+                    else:
+                        return await message.channel.send(content="Invalid class. Try again.", delete_after=10)
 
 
 async def setup(bot):
